Time: 38 ms (23.41%) | Memory: 44.4 MB (20.62%) - LeetSync

+++ Miscellaneous/115-distinct-subsequences/distinct-subsequences.cpp
@@ -12,7 +12,7 @@ public:
                 dp[i][j] = min(dp[i][j], (long long)INT_MAX);
             }
         }
-        return dp[n][m];
+        return (int)dp[n][m];
     }
 };
